fixes OCPBUGS-3909
The funcs that were writing systemd units to disk would skip the actual
file write step when the write is performed. However, the validation
funcs did not have this skip logic. This PR adds that missing logic as
well as a failing test case. In addition, this PR includes the following
opportunistic refactoring efforts:

- Hoist file and unit write funcs for reusability and places them in a
  separate file to make future refactoring easier.
- Minor refactors of some of the write funcs to make them more readable
  and more testable. In particular, they are reliant on global vars
  for determining which paths to use. These changes ensure that these
  funcs only write to a temp dir managed by the test framework.
- Refactors TestConfigDriftMonitor to use the newly hoisted file and
  unit write funcs.
- Use the default UID and GID and clear username for the test Ignition
  config. Permissions in CI are a bit more complicated than they are on
  an end-user machine since we're running in a container with an unknown
  username and unknown UID / GID. However, the defaults (-1 / -1) seem
  to work without issue as evidenced by
  writeFileAtomicallyWithDefaults() being able to write successfully.
